base class Translator
